study

+++ notnumbered/2057.py
@@ -5,12 +5,12 @@ if n == 0:
     n = -1
 res = 2432902008176640000
 for i in range(20,0,-1):
-    res = res/i
+    res = int(res/i)
     if n >= res:
         n = n - res
     
     
-if n == 0.0:
+if n == 0:
     print('YES')
 else:
     print('NO')
